Add manual token setting for local environment
Auth codes are immediately used when the auth flow redirects back to the site. This creates an issue where localhost environments cannot obtain a valid, unused auth code to exchange for an access token.

This provides a way to create valid auth on localhost by manually providing a valid token to the 'devCheckAuth' function. The token can be obtained from the cookie by using the console.